added movement podium interaction mechanics
instead of using ray casts to detect if player is in range of a podium for interactions, a trigger area has been set that the player must be touching for the upgrade panel to be shown.
if/when the player moves out of the interaction trigger area, the upgrades panel will hide and re-show the hud.